requirements.txt


# PyTorch and related
torch==2.7.0
torchinfo==1.8.0
transformers==4.46.2
datasets==3.1.0
sentence_transformers==3.0.1
argparse==1.4.0

# Data processing
numpy>=1.26.0
pandas==2.2.0
scipy==1.11.4
openai==1.84.0

# Visualization and TSA utilities
matplotlib==3.10.1
python-calamine==0.3.1
tslearn==0.6.3
ruptures==1.1.9
statsmodels==0.14.4
h5py==3.13.0


# R integration
rpy2==3.5.17

# Web
aiohttp==3.11.13


# TEdit
pyyaml==6.0.2
linear_attention_transformer==0.19.1
tensorboard==2.19.0
scikit-learn==1.3.2
